리펙토링: linux / windows upload
