+++ src/main/java/dev/valium/sweetmeme/module/processor/EnvProcessor.java
@@ -1,0 +1,9 @@
+package dev.valium.sweetmeme.module.processor;
+
+import dev.valium.sweetmeme.infra.config.FileConfig;
+
+public class EnvProcessor {
+    public static String getActiveProfile() {
+        return System.getProperty("spring.profiles.active");
+    }
+}
