+++ src/main/java/dev/valium/sweetmeme/module/processor/fileUploader/FileUploader.java
@@ -1,0 +1,7 @@
+package dev.valium.sweetmeme.module.processor.fileUploader;
+
+import org.springframework.web.multipart.MultipartFile;
+
+public interface FileUploader {
+    public String uploadFile(String path, MultipartFile file, boolean encode);
+}
